@@ -5,7 +5,7 @@ author: xionghao
 homepage: https://github.com/xionghaoo/flutter_core
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
